requirements.txt
aiohttp==3.6.2
aiosqlite==0.15.0
amqp==5.0.1
appdirs==1.4.4
asgiref==3.2.10
async-timeout==3.0.1
attrs==19.3.0
billiard==3.6.3.0
celery==5.0.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
colorama==0.4.3
discord==1.0.1
discord.py==1.4.1
distlib==0.3.1
Django==3.1
django-environ==0.4.5
django-filter==2.3.0
djangorestframework==3.11.1
filelock==3.0.12
httpie==2.2.0
idna==2.10
kombu==5.0.2
Markdown==3.2.2
multidict==4.7.6
nest-asyncio==1.4.0
prompt-toolkit==3.0.7
Pygments==2.6.1
pytz==2020.1
redis==3.5.3
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
typing-extensions==3.7.4.2
urllib3==1.25.10
vine==5.0.0
virtualenv==20.0.31
wcwidth==0.2.5
yarl==1.5.1